+++ CHANGELOG.md
@@ -12,6 +12,13 @@
 - Log message when execution of extension finished
 - Updated invalid ports message
 - User must have READ permission in `file_mode`
+- Configure residue using `--setup` flag
+
+### Config Changes
+- Unknown clients have been renamed to unmanaged clients
+- Known clients have been renamed to managed clients
+- Unknown loggers have been renamed to unmanaged loggers
+- Known loggers have been renamed to known loggers
 
 ### Core Fixes
 - Standard response code has quotes around the key